+Frame Dump option, +refactored duplicate code
* Frame Dump option:
It is now possible to dump raw frames for specified frame ranges. This
helps with debugging
* Refactored the "complicated" code to calculate window heights,
centroids, etc. Now window centroids are stored for both x and y value.
Before only the x-value was stored. The y-value needed to be calculated
from the index of the x-value which is dirty.